create different layout for logged in and not logged in
