+++ client/src/components/ContentLayout.jsx
@@ -1,0 +1,7 @@
+import styles from './ContentLayout.module.css';
+
+const ContentLayout = ({ children }) => {
+  return <div className={styles.container}>{children}</div>;
+};
+
+export default ContentLayout;
