+++ client/src/components/ContentLayout.module.css
@@ -1,0 +1,5 @@
+.container{
+    /* temporary styles */
+    background-color: #7f7f7f;
+    height: 500px;
+}
